+++ src/main/java/pages/Product.java
@@ -7,12 +7,14 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.PageFactory;
 
+import logs.CreateLog;
 import utilities.BaseClass;
 import utilities.WaitExpectedConditions;
 
 public class Product extends BaseClass
 {
 	WebDriver driver;
+	CreateLog log=new CreateLog();
 	public Product(WebDriver driver) {
 		super(driver);
 		this.driver = driver;
@@ -47,8 +49,9 @@ public class Product extends BaseClass
 				.findElement(By.xpath(
 						"//span[contains(text(),'" + finded_string + "')]/ancestor::a/div[2]/div[1]/div[2]/span/div"))
 				.getText();
-		System.out.println("Mobile Name:   " + name);
-		System.out.println("Mobile rating:   " + rating);
+		log.writeLog("Mobile Name:  ", name);
+		log.writeLog("Mobile rating:   ", rating);
+		
 
 	}
 }
