+++ src/main/java/utilities/Launch.java
@@ -5,14 +5,12 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.ie.InternetExplorerDriver;
 import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-
 import io.github.bonigarcia.wdm.WebDriverManager;
 
 public class Launch {
 	private WebDriver driver;
 
-	@BeforeMethod
+	
 	public void initWebDriver() {
 		switch (Constants.BROWSER) {
 		case "Chrome":
